Update to the last rust-nigthly.
Version of rustc: 0.11.0-pre-nightly (9f8149e185fe55751b8d8675021d2066249abe54 2014-06-25 18:51:21 +0000).

+++ src/window/window.rs
@@ -509,7 +509,7 @@ impl<'a> Window<'a> {
 
     fn update_viewport(&mut self, w: f32, h: f32) {
         // Update the viewport
-        verify!(gl::Scissor(0 as i32, 0 as i32, w as i32, h as i32));
+        verify!(gl::Scissor(0, 0, w as i32, h as i32));
         FramebufferManager::screen().resize(w, h);
         self.post_process_render_target.resize(w, h);
     }
